% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_weather}
\alias{daily_weather}
\title{Daily Weather Data}
\format{
matrix (ts object) with 4383 rows and 6 columns.
\describe{
\item{min_temp}{Minimum temperature of the day (celcius)}
\item{max_temp}{Maximum temperature of the day (celcius)}
\item{avg_temp}{Average temperature of the day (celcius)}
\item{avg_humidity}{Average humidity of the day (\%)}
\item{avg_rainfall}{Average rainfall of the day (mm)}
\item{sunshine}{Duration of sunshine of the day (hour)}
}
}
\source{
\url{https://dataonline.bmkg.go.id/}
}
\usage{
daily_weather
}
\description{
Daily weather data collected by BMKG in Citeko Meteorological Station, Bogor
Regency, Indonesia. The data collected are from Januari 1991 to December 2002.
}
\keyword{datasets}
