% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_outliers.R
\name{dummy_outliers}
\alias{dummy_outliers}
\title{Identifying outliers}
\usage{
dummy_outliers(init_data, sig_level = 0.05)
}
\arguments{
\item{init_data}{matrix (ts object), the initially imputed data with n rows
(observation) and k columns (variables). init_data cannot have
missing values.}

\item{sig_level}{numeric, significant level for Chi-Square test.}
}
\value{
A matrix (ts object), containing indexes of the outliers from each
variable. 1 means outlier while 0 means not outlier.
}
\description{
This function identifies outliers on each variable using forecast::outliers.
}
\examples{
dum = dummy_outliers(monthly_weather)
sum(dum) # 16 observations are considered outliers in `monthly_weather` data

}
