% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_imp.R
\name{init_imp}
\alias{init_imp}
\title{Initial imputation}
\usage{
init_imp(miss_data, init_method = "na_interpolation")
}
\arguments{
\item{miss_data}{matrix (ts object), missing data with n rows (observation) and k
columns (variables).}

\item{init_method}{character, initial imputation method (univariate) that will
be used. The univariate imputation methods are based on imputeTS package,
and must be one of the "na_interpolation", "na_kalman", "na_locf",
"na_ma", "na_mean", "na_replace", "na_seadec".}
}
\value{
A matrix with the same dimension as the \code{miss_data}, but with
temporarily imputed values.
}
\description{
VAR cannot be directly modeled on data with missing values, therefore, an
initial imputation is needed as the temporary data. These temporary data will
then be optimized using VAR-IM.
}
\examples{
# simultaneously imputing missing values on multiple variables
miss_mult = omit(monthly_weather, seed = 1, num_miss = 10, first_var = FALSE)
colSums(is.na(miss_mult)) # number of missing values
imp_mult = init_imp(miss_mult)
colSums(is.na(imp_mult)) # number of missing values after initial imputation

}
