% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_all_stationary.R
\name{is_all_stationary}
\alias{is_all_stationary}
\title{ADF test for all variables}
\usage{
is_all_stationary(init_data, crit_value = 0.05)
}
\arguments{
\item{init_data}{matrix (ts object), the initially imputed data with n rows
(observation) and k columns (variables). init_data cannot have
missing values.}

\item{crit_value}{numeric, critical value for ADF test.}
}
\value{
Logical, TRUE if all variable are stationary, FALSE if there is at
least one variable that is non-stationary.
}
\description{
ADF test for all variables on matrix with k variables.
}
\examples{
is_all_stationary(daily_weather)
try(is_all_stationary(monthly_weather))
try(is_all_stationary(monthly_weather, crit_value = 0.01))
}
