% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_weather}
\alias{monthly_weather}
\title{Monthly Weather Data}
\format{
matrix (ts object) with 144 rows and 6 columns.
\describe{
\item{min_temp}{Minimum temperature of the day (celcius)}
\item{max_temp}{Maximum temperature of the day (celcius)}
\item{avg_temp}{Average temperature of the day (celcius)}
\item{avg_humidity}{Average humidity of the day (\%)}
\item{avg_rainfall}{Average rainfall of the day (mm)}
\item{sunshine}{Duration of sunshine of the day (hour)}
}
}
\source{
\url{https://dataonline.bmkg.go.id/}
}
\usage{
monthly_weather
}
\description{
Monthly weather data collected by BMKG in Citeko Meteorological Station, Bogor
Regency, Indonesia. The data collected are from Januari 1991 to December 2002.
This monthly data was calculated as the average of "daily_weather" data within
a month.
}
\keyword{datasets}
