% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit.R
\name{omit}
\alias{omit}
\title{Randomly remove some data}
\usage{
omit(
  completed_data,
  seed,
  num_miss,
  max_lag = 20,
  consecutive = TRUE,
  first_var = TRUE,
  completely_miss = FALSE
)
}
\arguments{
\item{completed_data}{matrix (ts object), complete data with n rows
(observation) and k columns (variables).}

\item{seed}{integer, unique number to control the random removal.}

\item{num_miss}{integer, number of data that will be removed}

\item{max_lag}{integer, maximum lag for VAR-IM.}

\item{consecutive}{logical, should the data be removed consecutively?}

\item{first_var}{logical, should the data be removed on first variable only?}

\item{completely_miss}{logical, should the data be removed simultaneously
(all variables in the same observation are missing)?}
}
\value{
A matrix with same dimension as the \code{completed_data}, but with
\code{num_miss} missing values.
}
\description{
There are 6 removal schemes based on whether the data should be missing
consecutively, only on one variable or on multiple variable, as well as whether
an observation can have non-missing variable.
}
\examples{
# The dafault setting will remove `num_miss` observations from the first variable.
miss1 = omit(monthly_weather, seed = 1, num_miss = 10)
colSums(is.na(miss1)) # number of missing values of each variable

# to remove values on all variables (instead of the first variable only), use
# `first_var = FALSE`.
miss2 = omit(monthly_weather, seed = 1, num_miss = 10, first_var = FALSE)
colSums(is.na(miss2)) # number of missing values of each variable

# If `first_var = FALSE` and `completely_miss = TRUE`, all variables of the
# randomly chosen observations will have missing values.
miss3 = omit(monthly_weather, seed = 1, num_miss = 10, first_var = FALSE, completely_miss = TRUE)
colSums(is.na(miss3)) # number of missing values of each variable
miss_index = which(is.na(miss3), arr.ind = TRUE)
miss_index[order(miss_index[,"row"],decreasing=FALSE),] # index of the missing values

}
