% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARDim.R
\name{VARDim}
\alias{VARDim}
\title{VARD-IM}
\usage{
VARDim(
  miss_data,
  d,
  ic = "AIC",
  min_norm = 0.05,
  max_iter = 10,
  max_lag = 30,
  p = NULL,
  init_method = "na_interpolation",
  dummy = NULL,
  outliers = FALSE
)
}
\arguments{
\item{miss_data}{matrix (ts object), missing data with n rows (observation)
and k columns (variables).}

\item{d}{integer, differencing order.}

\item{ic}{character, information criteria for model selection. The value has
to be either "AIC", "HQ", or "SC".}

\item{min_norm}{numeric, threshold for the iteration of VAR modeling. The
smaller the min_norm is, the longer the imputation takes to complete.}

\item{max_iter}{integer, number of maximum iteration if min_norm is not met.}

\item{max_lag}{integer, maximum lag of VAR to be considered.}

\item{p}{integer, lag of the VAR model if needed to be set manually.}

\item{init_method}{character, initial imputation method (univariate) that will
be used. The univariate imputation methods are based on imputeTS package,
and must be one of the "na_interpolation", "na_kalman", "na_locf",
"na_ma", "na_mean", "na_replace", "na_seadec".}

\item{dummy}{matrix, exogen variables, varibales that are not included in the
predictive model but can be used to improve the overall model. The
matrix has to be in m x l dimension, where m is the number of
observation and l is the number of the dummy variables.}

\item{outliers}{logical, if the logical == TRUE, dummy_outliers function will
identify outlier on each variable.}
}
\value{
A list with 4 objects:
\enumerate{
\item \code{imputed_data}: matrix/ts object of the final imputed data,
\item \code{convergency}: matrix that details the norm/convergency of each iteration,
\item \code{convergent_model}: VAR model of the convergent/last iteration, and
\item \code{coefficient}: coefficient matrix of the last iteration
}
}
\description{
This function performs an imputation using Vector Autoregression Differencing
Imputation Method (VARD-IM), a variation of VAR-IM which uses Differencing.
This function is more suitable for non-stationary multivariate time series data.
}
\examples{
miss_data = omit(daily_weather, seed = 1, num_miss = 100)
colSums(is.na(miss_data))
# For data with 100 missing values, the most basic imputation that can be used are:
imp_data = VARDim(miss_data, d = 1)

# Since VARDim is intended to be used for non-stationary data that needs a
# differencing process, order of differencing `d` has to be input.

# For better imputation results, a related dummy variable(s) may be used.
# The following examples are just illustration and clearly are not related to
# the data at all.

# `dummy` properties can use a vector or a matrix with k columns (k variables).
d = 1 # order of differencing
num_row = nrow(daily_weather) - d
dummy_mat = rnorm(num_row,0,1)
imp_data_OneDummy = VARDim(miss_data, d = 1, dummy = dummy_mat)

# For dummy with multiple variables:
dummy_mats = cbind(rnorm(num_row,0,1), rnorm(num_row,1,4))
imp_data_MultipleDummy = VARDim(miss_data, d = 1, dummy = dummy_mats)

}
